Fix issue with access control module
