+++ modules/access-control/src/components/access-control-show.tsx
@@ -77,7 +77,8 @@ export const Show: FunctionComponent<PropsWithChildren<AccessControlShowInterfac
     const hasScopesToShow: boolean = useRequiredScopes(when);
     const hasScopesToHide: boolean = useRequiredScopes(notWhen);
 
-    let showOn: boolean = false;
+    // By default, show the component.
+    let showOn: boolean = true;
     let hideOn: boolean = false;
 
     // If `when` is defined, use the hook to check if the user has the required scopes.
